refactor(server): consolidate smallest handler modules to reduce file fragmentation
Merged extension-handlers.js (44 lines, 1 fn), web-task-handlers.js
(57 lines, 4 fns), and environment-handlers.js (120 lines, 4 fns)
into a single misc-handlers.js (189 lines, 9 fns). Reduces handlers/
from 10 files to 8 files with no behavior changes.

Updated ws-message-handlers.js import list and the one test that
imported environment-handlers directly.

Closes #2708
